+++ Dockerfile
@@ -4,4 +4,4 @@ RUN apk add --virtual .build-dependencies build-base libressl libffi-dev && \
   pip install -r requirements.txt && \
   apk del .build-dependencies
 ADD webtop .
-ENTRYPOINT python webtop/__init__.py
+ENTRYPOINT python webtop/cli.py
